Trying to calculate the critical cluster radius.
The expression, as given in Clouet, is very sensitive to the G_nuc expression provided. 

A better approach would be to see at what value of n the alpha and beta functions match, using some kind of binary search, Before doing this, we need values of the jump distance and jump frequency.